+++ src/card/behavior.rs
@@ -29,7 +29,6 @@ pub enum CardBehavior {
     Moat,
     Mulch,
     Negate,
-    NevinyrralsDisk,
     PillarOfFlame,
     Recall,
     SedgeTroll,